crawler_pricing_cms/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

awswrangler==2.20.0
boto3==1.24.28
gdown==4.6.4
pandas==1.4.4
tabula-py==2.7.0
